refactor(checker): split property access resolver helpers
