+++ crates/tsz-checker/src/types/property_access_type/mod.rs
@@ -2,11 +2,14 @@
 //! and expando function pattern detection.
 
 mod class_recovery;
+mod enum_namespace_access;
 mod helpers;
+mod identifier_resolution;
 mod imported_array_to_enum;
 pub(crate) mod known_globals;
 mod nullish_access;
 mod optional_chain_cache;
+mod optional_fast_path;
 mod partial_initializer;
 mod resolve;
